+++ client/src/modelView/DomainTypesView.ts
@@ -17,13 +17,15 @@ import { PddlTokenType } from '../../../common/src/PddlTokenizer';
 import { nodeToRange } from '../utils';
 import { DocumentInsetCodeLens, DocumentCodeLens } from './view';
 import { DomainView, DomainRendererOptions, DomainRenderer } from './DomainView';
+import { GraphViewData, NetworkEdge, NetworkNode } from './GraphViewData';
+import { DomainViewPanel } from './DomainViewPanel';
 
-const CONTENT = 'modelView';
+const CONTENT = path.join('views', 'modelView');
 
 const PDDL_DOMAIN_TYPES_PREVIEW_COMMAND = "pddl.domain.types.preview";
 const PDDL_DOMAIN_TYPES_INSET_COMMAND = "pddl.domain.types.inset";
 
-export class DomainTypesView extends DomainView<DomainTypesRendererOptions, DomainTypesViewData> implements CodeLensProvider {
+export class DomainTypesView extends DomainView<DomainTypesRendererOptions, GraphViewData> implements CodeLensProvider {
 
     constructor(context: ExtensionContext, codePddlWorkspace: CodePddlWorkspace) {
         super(context, codePddlWorkspace, new DomainTypesRenderer(), {
@@ -32,7 +34,7 @@ export class DomainTypesView extends DomainView<DomainTypesRendererOptions, Doma
             insetViewCommand: PDDL_DOMAIN_TYPES_INSET_COMMAND,
             insetHeight: 5,
             webviewType: 'domainTypesPreview',
-            webviewHtmlPath: 'problemObjectsView.html',
+            webviewHtmlPath: 'graphView.html',
             webviewOptions: {
                 enableFindWidget: true,
                 // enableCommandUris: true,
@@ -54,11 +56,16 @@ export class DomainTypesView extends DomainView<DomainTypesRendererOptions, Doma
         if (!domain) { return []; }
 
         let defineNode = domain.syntaxTree.getDefineNodeOrThrow();
-        let typesNode = defineNode.getFirstChildOrThrow(PddlTokenType.OpenBracketOperator, /\s*:types/i);
-        return [
-            new DocumentCodeLens(document, nodeToRange(document, typesNode)),
-            new DocumentInsetCodeLens(document, nodeToRange(document, typesNode), document.positionAt(typesNode.getStart()).line)
-        ];
+        let typesNode = defineNode.getFirstChild(PddlTokenType.OpenBracketOperator, /\s*:types/i);
+        if (typesNode) {
+            return [
+                new DocumentCodeLens(document, nodeToRange(document, typesNode)),
+                new DocumentInsetCodeLens(document, nodeToRange(document, typesNode), document.positionAt(typesNode.getStart()).line)
+            ];
+        }
+        else {
+            return [];
+        }
     }
 
     async resolveCodeLens(codeLens: CodeLens, token: CancellationToken): Promise<CodeLens> {
@@ -83,10 +90,15 @@ export class DomainTypesView extends DomainView<DomainTypesRendererOptions, Doma
     protected createPreviewPanelTitle(uri: Uri) {
         return `:types in '${path.basename(uri.fsPath)}'`;
     }
+
+    protected async handleOnLoad(panel: DomainViewPanel): Promise<boolean> {
+        await panel.postMessage('setInverted', { value: true });
+        return super.handleOnLoad(panel);
+    }
 }
 
-class DomainTypesRenderer implements DomainRenderer<DomainTypesRendererOptions, DomainTypesViewData> {
-    render(context: ExtensionContext, domain: DomainInfo, options: DomainTypesRendererOptions): DomainTypesViewData {
+class DomainTypesRenderer implements DomainRenderer<DomainTypesRendererOptions, GraphViewData> {
+    render(context: ExtensionContext, domain: DomainInfo, options: DomainTypesRendererOptions): GraphViewData {
         let renderer = new DomainTypesRendererDelegate(context, domain, options);
 
         return {
@@ -94,11 +106,6 @@ class DomainTypesRenderer implements DomainRenderer<DomainTypesRendererOptions, 
             relationships: renderer.getRelationships()
         };
     }
-}
-
-interface DomainTypesViewData {
-    nodes: NetworkNode[];
-    relationships: NetworkEdge[];
 }
 
 class DomainTypesRendererDelegate {
@@ -136,14 +143,3 @@ class DomainTypesRendererDelegate {
 
 interface DomainTypesRendererOptions extends DomainRendererOptions {
 }
-
-interface NetworkNode {
-    id: number;
-    label: string;
-}
-
-interface NetworkEdge {
-    from: number;
-    to: number;
-    label: string;
-}