+++ client/views/modelView/graphView.js
@@ -8,6 +8,16 @@ var networkData = {
 
 var network = null;
 
+var _inverted = false;
+const TOP_DOWN = "TOP_DOWN";
+const LEFT_RIGHT = "LEFT_RIGHT";
+var _layout = TOP_DOWN;
+var _settings = false;
+var _origData = {
+  nodes: [],
+  relationships: []
+};
+
 function initialize() {
   // create a network
   var container = document.getElementById("network");
@@ -38,18 +48,11 @@ function initialize() {
       hierarchical: {
         enabled: true,
         direction: "DU",
-        sortMethod: "directed",
-        levelSeparation: 77,
-        nodeSpacing: 17,
-        treeSpacing: 17,
-        parentCentralization: false
+        sortMethod: "directed"
       }
     },
     physics: {
-      enabled: true,
-      hierarchicalRepulsion: {
-        centralGravity: 0
-      },
+      enabled: false,
       minVelocity: 0.75,
       solver: "hierarchicalRepulsion"
     },
@@ -71,6 +74,7 @@ function initialize() {
   })
 
   if (!vscode) { populateWithTestData(); }
+  ensureLayout();
   onLoad();
 }
 
@@ -78,6 +82,12 @@ function handleMessage(message) {
   switch (message.command) {
     case 'updateContent':
       updateGraph(message.data);
+      break;
+    case 'setInverted':
+      setInverted(message.value);
+      break;
+    case 'setOptions':
+      setOptions(message.options);
       break;
     default:
       console.log("Unexpected message: " + message.command);
@@ -88,7 +98,7 @@ function handleMessage(message) {
 function populateWithTestData() {
   // for testing only
   updateGraph({
-    nodes: [{ id: 1, label: 'City' }, { id: 2, label: 'Town' }, { id: 3, label: 'Village' }, { id: 4, label: 'Capital' }],
+    nodes: [{ id: 1, label: 'City' }, { id: 2, label: 'Town nnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnn' }, { id: 3, label: 'Village' }, { id: 4, label: 'Capital' }],
     relationships: [{ from: 1, to: 2 }, { from: 2, to: 3 }, {from: 4, to: 2}]
   });
   setIsInset(true);
@@ -100,8 +110,9 @@ function clearNetwork() {
 }
 
 function updateGraph(data) {
+  _origData = data;
   clearNetwork();
-  nodes.add(data.nodes);
+  if (data.nodes) nodes.add(data.nodes);
   edges.add(data.relationships);
   network.fit({animation: true});
 }
@@ -115,19 +126,56 @@ function resize() {
   }
 }
 
+/**
+ * Sets the `inverted` fag. 
+ * @param {boolean} inverted should the direction of the graph layout be inverted?
+ */
+function setInverted(inverted) {
+  _inverted = inverted;
+  ensureLayout();
+}
+
+function ensureLayout() {
+  if (_layout === TOP_DOWN) {
+    topDown();
+  }
+  else if (_layout === LEFT_RIGHT) {
+    leftRight();
+  }
+}
+
 function topDown() {
-  setLayoutDirection('DU');
+  _layout = TOP_DOWN;
+  setLayoutDirection(_inverted ? 'DU' : 'UD');
 }
 
 function leftRight() {
-  setLayoutDirection('RL');
+  _layout = LEFT_RIGHT;
+  setLayoutDirection(_inverted ? 'RL' : 'LR');
 }
 
 function setLayoutDirection(direction) {
   network.setOptions({ layout: { hierarchical: { direction: direction } } });
-  postMessage({ command: 'layout', direction: direction });
+}
+
+function setOptions(options) {
+  network.setOptions(options);
 }
 
 function fit() {
   network.fit();
+}
+
+function toggleSettings() {
+  _settings = !_settings;
+  network.setOptions({ configure: _settings });
+  document.body.style.overflow = _settings ? 'scroll' : 'hidden';
+  if (_settings) {
+    var settingsElement = document.getElementsByClassName("vis-configuration-wrapper")[0];
+    if (settingsElement) settingsElement.scrollIntoView();
+  }
+}
+
+function reset() {
+  updateGraph(_origData);
 }