+++ client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # PDDL support - What's new?
 
+## [2.15.1] Fixes and upgrades
+
+- upgrade to vscode 1.40
+- upgrade to TypeScript 3.7.2
+- clean-up
+- fix for type hierarchy parsing that miss the " - object"
+- moving views to a dedicated 'views' folder
+- more robust action/variable parameter parsing
+- eagerly using the Elvis operator
+- constraint partial order view
+- network/graph settings UI
+- improved icons
+- types vs objects visualized in different color
+
 ## [2.15.0] PDDL Model Visualization
 
 ### Domain `:types` hierarchy and problem `:objects` and `:init` visualization